fix(core): Resolve comments
